+++ CHANGELOG.md
@@ -23,6 +23,9 @@ The changelog for **SwifterSwiftUI**. Also see the [releases](https://github.com
 - **Mutable Protocol**:
     - Adds a helper function to mutate a properties and help implement builder pattern. Very useful to create methods to modify SwiftUI views. [#6](https://github.com/asam139/Steps/pull/6) by [asam139](https://github.com/asam139)
 
+- **View**:
+    - `frame(size:)`.  [#6](https://github.com/asam139/Steps/pull/6) by [asam139](https://github.com/asam139)
+
 ## [v0.4.2](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.4.2)
 
 - **Utils**: